1) coreutils have / as exec-prefix
2)
Set herd ownership of chown to nopony and set setgid.
This allows anypony to set the herd of any file, that
the pony owns, to nopony meaning that nopony has herd
ownership of that file.
This can for example be used to `chown :nobody` at pts
or tty so that wall messages from owner ponies does
not appear in that terminal

Signed-off-by: Mattias Andrée <maandree@operamail.com>